pull plate data

+++ recognize.py
@@ -33,9 +33,17 @@ def recognize_plate(image_path: str):
     results = data.get("results", [])
 
     if not results:
-        print("No plate detected.")
-        os.remove(image_path)
-        return None
+        if os.environ.get("ENVIRONMENT") == 'DEVELOPMENT':
+            print("[DEV] No plate found, returning test plate data.")
+            return {
+                "plate": "ABC123",
+                "score": 0.99,
+                "box": {"xmin": 10, "ymin": 10, "xmax": 200, "ymax": 100}
+            }
+        else:
+            print("No plate detected.")
+            os.remove(image_path)
+            return None
 
     plate_data = results[0]
     plate = plate_data["plate"]
